==> notebooks/05_modeling.ipynb <==
# Modeling & Evaluation

# lets import needed libraries

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score

# Load Dataset

product_monthly = pd.read_csv("../data/processed/feature_engineered_sales.csv", parse_dates=['Month'])
product_monthly.head()


### Train-Test Split (Time-Aware, Across Products)

# Here lets define features and target variable

X = product_monthly.drop(columns=['Sales', 'Month'])
y = product_monthly['Sales']

# here we encode the product names

X = pd.get_dummies(X, columns=['Product'], drop_first=True)

# Time- based split

split_date = product_monthly['Month'].quantile(0.8)

train_idx = product_monthly['Month'] <= split_date
test_idx = product_monthly['Month'] > split_date

X_train, X_test = X[train_idx], X[test_idx]
y_train, y_test = y[train_idx], y[test_idx]


# Let's train Models

lr_model = LinearRegression()
lr_model.fit(X_train, y_train)

rf_model = RandomForestRegressor(random_state=42)
rf_model.fit(X_train, y_train)


# predictions

lr_preds = lr_model.predict(X_test)
rf_preds = rf_model.predict(X_test)

# Lets evaluate the models

lr_mae = mean_absolute_error(y_test, lr_preds)
lr_rmse = np.sqrt(mean_squared_error(y_test, lr_preds))
lr_r2 = r2_score(y_test, lr_preds)

rf_mae = mean_absolute_error(y_test, rf_preds)
rf_rmse = np.sqrt(mean_squared_error(y_test, rf_preds))
rf_r2 = r2_score(y_test, rf_preds)


# Let's save predictions and metrics

results_df = pd.DataFrame({
    "Month": product_monthly.loc[test_idx, "Month"],
    "Actual Sales": y_test,
    "LR_Predicted": lr_preds,
    "RF_Predicted": rf_preds
})

results_df.to_csv("../data/processed/model_predictions.csv", index=False)

metrics_df = pd.DataFrame({
    "Model": ["Linear Regression", "Random Forest"],
    "MAE": [lr_mae, rf_mae],
    "RMSE": [lr_rmse, rf_rmse],
    "R²": [lr_r2, rf_r2]
})

metrics_df.to_csv("../data/processed/model_metrics.csv", index=False)

metrics_df
